all frontend works. CSS not working properly. Mentor backend not working yet.

+++ frontend/src/admin-dashboard/MentorList.jsx
@@ -18,7 +18,7 @@ const db = getFirestore(app);
 const Mentors = () => {
   const theme = useTheme();
   const colors = tokens(theme.palette.mode);
-  const [pendingMentors, setPendingMentors] = useState([]);
+  const [mentors, setMentors] = useState([]);
 
   useEffect(() => {
     const unsub = onSnapshot(collection(db, "mentors"), (snapshot) => {
@@ -27,19 +27,28 @@ const Mentors = () => {
         name: doc.data().name || "N/A",
         email: doc.data().email || "N/A",
         university: doc.data().university || "N/A",
+        industry: doc.data().industry || "N/A",
         major: doc.data().major || "N/A",
+        skills: doc.data().skills || "N/A",
       }));
-      setPendingMentors(data);
+      setMentors(data);
     });
     return () => unsub();
   }, []);
 
   const handleDelete = async (id) => {
+    const mentor = mentors.find((m) => m.id === id);
+    const confirmDelete = window.confirm(
+      `Are you sure you want to delete mentor '${mentor.name}'?`
+    );
+
+    if (!confirmDelete) return;
+
     try {
       await deleteDoc(doc(db, "mentors", id));
-      alert("Mentor deleted!");
+      setMentors((prev) => prev.filter((m) => m.id !== id));
     } catch (error) {
-      console.error("Error deleting Mentor:", error);
+      console.error("Error deleting mentor:", error);
     }
   };
 
@@ -66,12 +75,27 @@ const Mentors = () => {
       align: "center",
     },
     {
+      field: "industry",
+      headerName: "Industry",
+      flex: 1,
+      headerAlign: "center",
+      align: "center",
+    },
+    {
       field: "major",
       headerName: "Major",
       flex: 1,
       headerAlign: "center",
       align: "center",
     },
+    {
+      field: "skills",
+      headerName: "Skills",
+      flex: 1,
+      headerAlign: "center",
+      align: "center",
+    },
+
     {
       field: "actions",
       headerName: "Actions",
@@ -125,7 +149,7 @@ const Mentors = () => {
           },
         }}
       >
-        <DataGrid rows={pendingMentors} columns={columns} />
+        <DataGrid rows={mentors} columns={columns} />
       </Box>
     </Box>
   );
